+++ src/test/java/api/APITestSuite.java
@@ -36,6 +36,7 @@ import java.util.function.Consumer;
 @RunWith(Suite.class)
 
 @Suite.SuiteClasses({
+  CreateByBarcodeTests.class,
   LoanAPITests.class,
   LoanAPILocationTests.class,
   LoanAPITitleTests.class,