+++ src/test/java/api/loans/CreateByBarcodeTests.java
@@ -1,0 +1,37 @@
+package api.loans;
+
+import io.vertx.core.json.JsonObject;
+import api.support.APITests;
+import org.folio.circulation.support.http.client.IndividualResource;
+import org.junit.Test;
+
+import java.net.MalformedURLException;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+public class CreateByBarcodeTests extends APITests {
+
+  @Test
+  public void canCreateALoanUsingItemAndUserBarcode()
+    throws InterruptedException,
+    MalformedURLException,
+    TimeoutException,
+    ExecutionException {
+
+    final IndividualResource smallAngryPlanet = itemsFixture.basedUponSmallAngryPlanet();
+    final IndividualResource steve = usersFixture.steve();
+
+    final IndividualResource response = loansFixture.checkOutByBarcode(smallAngryPlanet, steve);
+
+    final JsonObject loan = response.getJson();
+
+    assertThat(loan.getString("id"), is(notNullValue()));
+
+    assertThat("status is not open",
+      loan.getJsonObject("status").getString("name"), is("Open"));
+  }
+}